mini_projet/personne_service/requirements.txt
flask
flask-SQLAlchemy
Pyjwt